start dockerize

+++ src/main/java/com/abdellah/blog/repository/PostRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.abdellah.blog.model.Post;
 
-public interface PostRepository extends JpaRepository<Post, Integer> {
+public interface PostRepository extends JpaRepository<Post, Long> {
   // is it important to implement this interface? -> no, it's already implemented in the spring framework
 }
+
+// how to change the package name in vs code? -< ctrl + shift + p -> java: change package name
